fix: 댓글 조회 시 대댓글 갯수 반환하도록 수정

+++ core/src/main/java/com/port90/core/comment/infrastructure/impl/repository/persistence/CommentJpaRepository.java
@@ -27,5 +27,5 @@ public interface CommentJpaRepository extends JpaRepository<CommentEntity, Long>
     @Query("select c from Comment c where c.id = :commentId")
     Optional<CommentEntity> findByIdWithOptimisticLock(@RequestParam("commentId") Long commentId);
 
-    int countByParentId(Long parentId);
+    long countByParentId(Long parentId);
 }
